Complete Bug Audit and TDLib Compatibility Fixes
This commit addresses a comprehensive list of bugs and compatibility issues found during a system-wide audit of the Tele-Drive application.

Key changes include:
- TdLib Compatibility: Updated all TdApi object initializations to use the no-arg constructor + manual field assignment pattern, strictly adhering to the signatures in com.github.tdlibx:td:1.8.56.
- Durable Metadata Sync: Overhauled the synchronization logic in TeleDriveRepository to use an 'upsert' approach instead of wiping the local database. This ensures that local state is preserved and reconciled with Telegram Journal events, preventing data loss.
- Atomic Journaling: Added session guards to all journal-driven operations in ViewModels and Workers. Updated UploadWorker to wait for a stable Telegram message ID before recording metadata events.
- Metadata Integrity: Fixed JSON serialization in MetadataHelper to manually map data types (Long, Boolean, null), ensuring payload integrity across all supported Android versions.
- UI Stability: Resolved Material 3 compilation errors and missing imports in FileExplorerScreen and SettingsScreen. Consolidated @OptIn annotations for clean build execution.

Verified successful compilation with ./gradlew compileDebugKotlin. Binary artifacts and temporary scripts have been removed.

+++ app/src/main/java/com/example/teledrive/viewmodel/LoginViewModel.kt
@@ -159,7 +159,7 @@ class LoginViewModel(
 
                 repository.saveSession(
                     UserSession(
-                        telegramUserId = me.id,
+                        telegramUserId = me.id.toLong(),
                         phoneNumber = phoneNumber,
                         username = me.usernames?.activeUsernames?.firstOrNull() ?: "",
                         firstName = me.firstName ?: "",
